+++ StopRugRemoval/StopRugRemoval/ModEntry.cs
@@ -1,6 +1,7 @@
 ﻿using System.Reflection;
 using AtraShared.Integrations;
 using AtraShared.MigrationManager;
+using AtraShared.Utils;
 using AtraShared.Utils.Extensions;
 using HarmonyLib;
 using StardewModdingAPI.Events;
@@ -17,20 +18,15 @@ namespace StopRugRemoval;
 /// </summary>
 public class ModEntry : Mod
 {
-    private static readonly Lazy<IReflectedField<Multiplayer>> MultiplayerLazy = new(() => ReflectionHelper!.GetField<Multiplayer>(typeof(Game1), "multiplayer"));
-
     [SuppressMessage("StyleCop.CSharp.NamingRules", "SA1306:Field names should begin with lower-case letter", Justification = "Reviewed.")]
     private static GMCMHelper? GMCM = null;
 
     private MigrationManager? migrator;
 
-    
-
     /// <summary>
-    /// Gets Game1.multiplayer.
+    /// Gets a function that gets Game1.multiplayer.
     /// </summary>
-    /// <remarks>This still requires reflection and is likely slow.</remarks>
-    internal static Multiplayer Multiplayer => MultiplayerLazy.Value.GetValue();
+    internal static Func<Multiplayer> Multiplayer => MultiplayerHelpers.GetMultiplayer;
 
     // the following three properties are set in the entry method, which is approximately as close as I can get to the constructor anyways.
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.