+++ StopRugRemoval/StopRugRemoval/HarmonyPatches/Niceties/CrabPotPatches.cs
@@ -13,7 +13,8 @@ internal static class CrabPotPatches
     /// Surpresses drawing crabpots when an event is up.
     /// </summary>
     /// <returns>True to continue (and draw), false to surpress drawing.</returns>
+    [HarmonyPrefix]
     [HarmonyPatch(nameof(CrabPot.draw))]
     private static bool PrefixDraw()
-        => !ModEntry.Config.HideCrabPots || !Game1.eventUp;
+        => !ModEntry.Config.HideCrabPots || !(Game1.eventUp || Game1.isFestival());
 }